Activate build-scans, triggered by an environment variable
`--scan` will send the build scan to the public Gradle servers as usual.

On the other hand, setting up an environment variable named which points
to a Gradle Enterprise server will trigger automatic publications to
that server, regardless of the --scan parameter.

The variable must be named `GRADLE_ENTERPRISE_URL`.